# docs/openapi.yaml
openapi: 3.1.0

info:
  version: "0.1"
  title: Cutting edge localization Service API
  description: |
    Enter here a whole verbose description of the API in markdown format
    # Introduction
    Lorem ipsume
    # Some Definitions
    Lorem ipsume
  contact:
    name: Out Team name
    email: team@email.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: /api/docs/assets/translate-logo.png
    altText: Translation app logo
  tags:
    - name: Organization
      description: |
        Markdown description of the organization tag
    - name: Project
      description: |
        Markdown description of the project tag
    - name: Translation
      description: |
        Markdown description of the translation tag
    - name: Comment
      description: |
        Markdown description of the comment tag
  x-tagGroups:
    - name: Translation group
      tags:
        - Translation

paths:
  /api/projects:
    post:
      summary: Create a new Project
      description: |
        This endpoint allows you to create a new project that will hold all its translations
      operationId: postProject
      requestBody:
        description: Required data to create a new Project
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequestBody"
      responses:
        201:
          $ref: "#/components/responses/CreateProjectResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Project
  /api/projects/{projectName}:
    put:
      summary: Edit a Project
      description: |
        This endpoint allows you to to edit and modify a project and its information
      operationId: putProject
      parameters:
        - $ref: "#/components/parameters/ProjectNameParam"
      requestBody:
        description: Required data to edit the details of a project
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditProjectRequestBody"
      responses:
        201:
          $ref: "#/components/responses/EditProjectResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Project
  /api/translations/{projectName}:
    get:
      summary: Get Project translations
      description: |
        This endpoint allows you to get all the translations for a given Project.<br>
        You can also filter by specifying only the languages that you want.
      operationId: getProjectTranslations
      parameters:
        - $ref: "#/components/parameters/ProjectNameParam"
        - $ref: "#/components/parameters/GetProjectTranslationsLanguagesParam"
      responses:
        200:
          $ref: "#/components/responses/ProjectTranslationsGetResponse"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Translation
    post:
      summary: Create Project translations
      description: |
        This endpoint allows you to to add a translation for a specific project
      operationId: postProjectTranslation
      parameters:
        - $ref: "#/components/parameters/ProjectNameParam"
      requestBody:
        description: Required data to create a new translation with a key, for the project
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectTranslationRequestBody"
      responses:
        201:
          $ref: "#/components/responses/CreateTranslationResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Translation
    delete:
      summary: Delete Project translations
      description: This endpoint allows you to delete all the translations of a Project for a given "key"
      operationId: deleteProjectTranslations
      parameters:
        - $ref: "#/components/parameters/ProjectNameParam"
      requestBody:
        description: Required data to delete the translations of a project
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteProjectTranslationRequestBody"
      responses:
        204:
          $ref: "#/components/responses/ProjectTranslationsDeleteResponse"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Translation
  /api/translations/{projectName}/{translationId}:
    put:
      summary: Edit Project translations
      description: |
        This endpoint allows you to to edit and modify a translation for a specific project
      operationId: putProjectTranslation
      parameters:
        - $ref: "#/components/parameters/ProjectNameParam"
        - $ref: "#/components/parameters/TranslationIdParam"
      requestBody:
        description: Required data to edit a translation for the project
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditProjectTranslationRequestBody"
      responses:
        201:
          $ref: "#/components/responses/EditTranslationResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Translation
  /api/translation-comments/{translationId}:
    get:
      summary: Get Comments for a specific key
      description: |
        This endpoint list all the comments for a specific translation key in a project
      operationId: getTranslationComments
      parameters:
        - $ref: "#/components/parameters/TranslationIdParam"
      responses:
        200:
          $ref: "#/components/responses/TranslationCommentsGetResponse"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Comment
    post:
      summary: Create a comment for a translation key in a project
      description: |
        This endpoint allows you to add a comment to an existing translation key in a project
      operationId: postComment
      parameters:
        - $ref: "#/components/parameters/TranslationIdParam"
      requestBody:
        description: Required data to create a new Comment
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTranslationCommentRequestBody"
      responses:
        201:
          $ref: "#/components/responses/CreateTranslationCommentResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Comment
components:
  parameters:
    ProjectNameParam:
      name: projectName
      description: The name of the project.
      in: path
      required: true
      schema:
        type: string
      example:
        - traduction-project-1
    TranslationIdParam:
      name: translationId
      description: The id of the project's translation
      in: path
      required: true
      schema:
        type: string
      example:
        - 1234-4567-8901-2345
    GetProjectTranslationsLanguagesParam:
      name: languages
      description: A list of Language codes
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      example: |
        For English and French, use [ en, fr ] - Or for  Portugese, German and Spanish, use [ pt, de, es ]
  responses:
    ProjectTranslationsGetResponse:
      description: Get translations for a given project
      content:
        application/json:
          schema:
            type: object
            required:
              - translations
              - total
            properties:
              translations:
                description: The list of translations
                type: array
                items:
                  $ref: "#/components/schemas/TranslationKeyDTO"
              total:
                description: The total amount of translations
                type: integer
          examples:
            Successful response:
              value:
                translations:
                  - key: home
                    languages:
                      - code: en
                        text: home
                      - code: pt
                        text: casa
                  - key: contact
                    languages:
                      - code: en
                        tect: contact
                      - code: pt
                        tect: contato
                  - key: about us
                    languages:
                      - code: en
                        tect: about us
                      - code: pt
                        tect: sobre nós
                total: 3
    CreateProjectResponse:
      description: Response of a created project.
    EditProjectResponse:
      description: Response of an edited project.
    CreateTranslationResponse:
      description: Response of a created translation.
    EditTranslationResponse:
      description: Response of an edited translation.
#      allOf:
#        - $ref: "#/components/schemas/CreateTranslationResponse"
    ProjectTranslationsDeleteResponse:
      description: Response of a deleted translation key.
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: BadRequest. Returned when the payload is invalid.
    TranslationCommentsGetResponse:
      description: Get the comments for a given translation key in a project
      content:
        application/json:
          schema:
            type: object
            required:
              - comments
              - total
            properties:
              translations:
                description: The list of comments
                type: array
                items:
                  $ref: "#/components/schemas/CommentDTO"
              total:
                description: The total amount of comments for the translation key
                type: integer
          examples:
            Successful response:
              value:
                comments:
                  - userId: 25
                    text: What do you think about adding the italian translation?
                    createdAt: 2022-02-25T22:05:54.973Z
                  - userId: 47
                    text: I don't think. I'm having a coffee
                    createdAt: 2022-02-26T22:10:14.973Z
                total: 2
    CreateTranslationCommentResponse:
      description: Response of a created translation key's comment.
    NotFound:
      description: Returned when the requested ressource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
    Conflict:
      description: Conflict. Returned when resource cannot be created because of business rules violation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    TranslationKeyDTO:
      type: object
      required:
        - id
        - key
        - languages
      properties:
        id:
          type: string
        key:
          type: string
          maxLength: 20
          minLength: 1
        languages:
          type: array
          items:
            $ref: "#/components/schemas/TranslationDTO"
    TranslationDTO:
      type: object
      required:
        - code
        - text
      properties:
        code:
          type: string
        text:
          type: string
    CommentDTO:
      type: object
      required:
        - userId
        - text
        - createdAt
      properties:
        userId:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateProjectRequestBody:
      description: The payload body to create a new project
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 20
          minLength: 1
    EditProjectRequestBody:
      allOf:
        - $ref: "#/components/schemas/CreateProjectRequestBody"
    CreateProjectTranslationRequestBody:
      description: The payload body to create a translation
      type: object
      required:
        - key
        - code
        - text
      properties:
        key:
          type: string
        code:
          type: string
        text:
          type: string
    EditProjectTranslationRequestBody:
      description: The payload body to edit a translation
      allOf:
        - $ref: "#/components/schemas/CreateProjectTranslationRequestBody"
    DeleteProjectTranslationRequestBody:
      description: The payload body to delete translations for a key
      type: object
      required:
        - key
      properties:
        key:
          type: string
    CreateTranslationCommentRequestBody:
      description: The payload body to create a comment for a translation key
      type: object
      required:
        - userId
        - text
      properties:
        userId:
          type: string
        text:
          type: string
    Error:
      description: Standard Error response.
      type: object
      required:
        - error_type
        - name
        - message
      properties:
        error_type:
          type: string
#          description: "example: `UNAUTHORIZED`."
        name:
          type: string
          description: Error name.
        message:
          type: string
          description: The detailed reason of the error.
    NotFoundError:
      description: Ressource not found Error response.
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required:
             - error_type
          properties:
            error_type:
              type: string
              description: "example `The translations not found for this project`"
#  securitySchemes: